requirements.txt
django-debug-toolbar
django-imagekit==3.3
dj-database-url==0.4.1
Django==1.9.7
gunicorn==19.6.0
phonenumbers==7.4.4
pillow==3.3.0
psycopg2==2.6.1
requests==2.10.0
uk-postcode-utils==0.1.0
whitenoise==3.2
